+++ NEWS.md
@@ -1,5 +1,13 @@
 # qol 1.3.3 - DEVELOPMENT
 
+### New functions
+
+* `ifelse_multi()`: A variation of the `ifelse()` function which can handle multiple conditions in one go. The function takes in the conditions as characters. The conditions are parsed before evaluation to enable SAS like writing styles. (20.06.2026)
+
+### New functionality
+
+* `if.()`, `else_if.()`: Are now also able to use the new writing style with conditions as characters introduced by `ifelse_multi()`. (20.06.2026)
+
 ### Optimization
 
 * `summarise_plus()`: Brought percentiles up to more speed. (17.06.2026)